hotfix
Hotfix;

Passed the "bot" argument over to all commands for extended availability

Made sure the message-logs channel is created under the pre-existsing "Bot logs" category, if it exists for some reason or coincidence. If it doesn't exist it creates the category and the channel under that.

Added guildDelete.js so the bots deletes the database document for the guild it was just kicked from

Added join.js so bot can join voice channels